move prlib_intern to prpriv.h
